pyproject.toml
[tool.poetry]
name = "sdss-cherno"
version = "0.4.1a0"
description = "SDSS guider actor"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/cherno"
repository = "https://github.com/sdss/cherno"
documentation = "https://sdss-cherno.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "cherno" }
]
include = ["cherno/etc"]

[tool.poetry.scripts]
cherno = "cherno.__main__:cherno"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
sdsstools = ">=0.5.0"
sdss-clu = "^1.8.0"
astropy = "^5.0.0"
click-default-group = "^1.2.2"
sdss-coordio = "^1.5.1"
pandas = "^1.3.4"
photutils = "^1.3.0"
tables = ">=3.6.1"
simple-pid = "^1.0.1"
sqlalchemy = "^1.4.45"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
ipython = ">=8.0.0rc1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
Sphinx = {version=">=3.0.0"}
black = {version=">=20.8b1", allow-prereleases=true}
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
pandas-stubs = ">=1.5.0"

[tool.black]
line-length = 88
target-version = ['py311']
fast = true

[tool.isort]
profile = "black"
sections = ["FUTURE", "STDLIB", "TYPING", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "cherno"
known_typing = ["typing"]
known_sdss = ["clu", "sdsstools", "coordio", "fps_calibrations"]
lines_after_imports = 2

[tool.pytest.ini_options]
addopts = "--cov cherno --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
include = ["cherno/*"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
